// server/data/models/integrityOrderDetails.ts
import z from 'zod'

export type IntegrityOrderDetails = z.infer<typeof IntegrityOrderDetails>
export const IntegrityOrderDetails = z.object({
  specials: z.string(),
  legacySubjectId: z.string(),
  legacyOrderId: z.string().nullish(),
  firstName: z.string().nullish(),
  lastName: z.string().nullish(),
  alias: z.string().nullish(),
  dateOfBirth: z.string().nullish(),
  adultOrChild: z.string().nullish(),
  sex: z.string().nullish(),
  contact: z.string().nullish(),
  primaryAddressLine1: z.string().nullish(),
  primaryAddressLine2: z.string().nullish(),
  primaryAddressLine3: z.string().nullish(),
  primaryAddressPostCode: z.string().nullish(),
  phoneOrMobileNumber: z.string().nullish(),
  ppo: z.string().nullish(),
  mappa: z.string().nullish(),
  technicalBail: z.string().nullish(),
  manualRisk: z.string().nullish(),
  offenceRisk: z.boolean(),
  postCodeRisk: z.string().nullish(),
  falseLimbRisk: z.string().nullish(),
  migratedRisk: z.string().nullish(),
  rangeRisk: z.string().nullish(),
  reportRisk: z.string().nullish(),
  orderStartDate: z.string().nullish(),
  orderEndDate: z.string().nullish(),
  orderType: z.string().nullish(),
  orderTypeDescription: z.string().nullish(),
  orderTypeDetail: z.string().nullish(),
  wearingWristPid: z.string().nullish(),
  notifyingOrganisationDetailsName: z.string().nullish(),
  responsibleOrganisation: z.string().nullish(),
  responsibleOrganisationDetailsRegion: z.string().nullish(),
})